requirements.txt
openai~=1.14.1
python-telegram-bot~=21.1.1
pytesseract~=0.3.10
Pillow~=10.1.0